third_party/blink/web_tests/platform/generic/virtual/webrtc-wpt-plan-b/external/wpt/webrtc/legacy/onaddstream.https-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_true: Transceiver is set on event expected true got false
FAIL Check onaddstream assert_equals: expected "[{streams:[{}]},{streams:[{}]}]" but got "[]"
Harness: the test ran to completion.
